Conditional compilation of blackbox code.
