+++ src/assets/browser/app/src/controller/global/booting.rs
@@ -17,9 +17,13 @@ use controller::input::{
 };
 use controller::global::{ AppRunner, App, GlobalWeak, Global };
 use data::{ BackendConfigBootstrap, HttpManager, BackendConfig };
+
+use data::blackbox::BlackBoxDriver;
+#[cfg(any(not(deploy),console))]
 use data::blackbox::{ 
-    BlackBoxDriver, BlackBoxDriverImpl, HttpBlackBoxDriverImpl,
+    BlackBoxDriverImpl, HttpBlackBoxDriverImpl,
     NullBlackBoxDriverImpl };
+    
 use debug::{ DebugBling, create_interactors };
 use dom::{ domutil, Bling, NoBling };
 use dom::event::{ EventListener, Target, EventData, EventType, EventControl, ICustomEvent };
@@ -99,6 +103,23 @@ impl Booting {
         out
     }
     
+    #[cfg(any(not(deploy),console))]
+    fn make_blackbox(&self, debug_url: &Option<String>) -> BlackBoxDriver {
+        let reporter_driver : Box<BlackBoxDriverImpl> = if debug_url.is_some() {
+            let debug_url = self.config_url.join(&debug_url.as_ref().unwrap()).ok().unwrap();
+            console!("debug-url {:?}",debug_url);
+            Box::new(HttpBlackBoxDriverImpl::new(&self.http_manager,&debug_url))
+        } else {
+            Box::new(NullBlackBoxDriverImpl::new())
+        };
+        BlackBoxDriver::new(reporter_driver)
+    }
+    
+    #[cfg(all(deploy,not(console)))]
+    fn make_blackbox(&self, debug_url: &Option<String>) -> BlackBoxDriver {
+        BlackBoxDriver::new()
+    }
+    
     pub fn boot(&mut self, config: &BackendConfig) {
         console!("bootstrapping");
         let mut global = self.global.clone();
@@ -108,14 +129,7 @@ impl Booting {
             Box::new(NoBling::new())
         };
         let debug_url = config.get_debug_url();
-        let reporter_driver : Box<BlackBoxDriverImpl> = if debug_url.is_some() {
-            let debug_url = self.config_url.join(&debug_url.as_ref().unwrap()).ok().unwrap();
-            console!("debug-url {:?}",debug_url);
-            Box::new(HttpBlackBoxDriverImpl::new(&self.http_manager,&debug_url))
-        } else {
-            Box::new(NullBlackBoxDriverImpl::new())
-        };
-        let blackbox = BlackBoxDriver::new(reporter_driver);
+        let blackbox = self.make_blackbox(debug_url);
         let ar = AppRunner::new(
             &GlobalWeak::new(&global),&self.http_manager,
             &self.el,bling,&self.config_url,config,
