+++ src/assets/browser/app/src/data/blackbox/blackboxdriver.rs
@@ -1,0 +1,20 @@
+use std::sync::{ Arc, Mutex };
+
+use super::BlackBoxState;
+
+pub trait BlackBoxDriverImpl {
+    fn tick(&mut self, state: &mut BlackBoxState, t: f64) {}    
+}
+
+#[derive(Clone)]
+pub struct BlackBoxDriver(Arc<Mutex<Box<BlackBoxDriverImpl>>>);
+
+impl BlackBoxDriver {
+    pub fn new(driver: Box<BlackBoxDriverImpl>) -> BlackBoxDriver {
+        BlackBoxDriver(Arc::new(Mutex::new(driver)))
+    }
+    
+    pub fn tick(&mut self, state: &mut BlackBoxState, t: f64) {
+        self.0.lock().unwrap().tick(state,t);
+    }
+}
