+++ src/assets/browser/app/src/data/blackbox/mod.rs
@@ -1,3 +1,9 @@
+mod blackboxdriver;
+
+pub use self::blackboxdriver::{
+    BlackBoxDriverImpl, BlackBoxDriver,
+};
+
 mod bbreportstream;
 mod blackbox;
 mod blackboxstate;
@@ -6,7 +12,6 @@ mod nullblackboxdriver;
 
 pub use self::bbreportstream::BlackBoxReportStream;
 pub use self::blackbox::{ 
-    BlackBoxDriverImpl, BlackBoxDriver,
     blackbox_report, blackbox_push, blackbox_pop,
     blackbox_tick, blackbox_elapsed, blackbox_metronome,
     blackbox_is_enabled
