+++ src/assets/browser/app/src/data/blackbox/stubbbdriver.rs
@@ -1,0 +1,7 @@
+pub struct BlackBoxDriver();
+
+impl BlackBoxDriver {
+    pub fn new() -> BlackBoxDriver {
+        BlackBoxDriver()
+    }    
+}
