+++ src/assets/browser/app/src/data/mod.rs
@@ -1,13 +1,21 @@
 mod backendconfig;
 mod backendconfigbootstrap;
 mod backendstickmanager;
-pub mod blackbox;
 mod httpmanager;
 mod httpxferclerk;
 mod psychic;
 mod xfercache;
 mod xferclerk;
 mod xferrequest;
+
+#[cfg(any(not(deploy),console))]
+pub mod blackbox;
+
+#[cfg(all(deploy,not(console)))]
+pub mod blackbox {
+    mod stubbbdriver;
+    pub use self::stubbbdriver::BlackBoxDriver;
+}
 
 pub use self::backendconfig::{ BackendConfig, BackendBytecode };
 pub use self::backendconfigbootstrap::{ BackendConfigBootstrap };
